FixRoomRepository

+++ HotelManagementProject.Repository/Repositories/RoomRepository.cs
@@ -2,10 +2,11 @@
 using HotelManagementProject.Domain;
 using HotelManagementProject.Domain.Dtos;
 using HotelManagementProject.Domain.Entites;
+using HotelManagementProject.Domain.Intefacies;
 using Microsoft.EntityFrameworkCore;
 
 namespace HotelManagementProject.Repository.Repositories;
-public class RoomRepository
+public class RoomRepository : IRoomRepository
 {
     private readonly AppDbContext _context;
     public async Task<bool> AddRoomAsync(Room Room)
@@ -16,13 +17,17 @@ public class RoomRepository
 
     public async Task<bool> DeleteAsync(Guid id)
     {
-        var room = await GetByIdAsync(id);
+        var room = await GetRoomAsync(id);
+
+        if (room == null)
+            return false;
+
         _context.Rooms.Remove(room);
 
         return await _context.SaveChangesAsync() > 0;
     }
 
-    public async Task<Room> GetByIdAsync(Guid id)
+    public async Task<Room?> GetRoomAsync(Guid id)
     {
         return await _context.Rooms.FindAsync(id);
     }
@@ -34,12 +39,16 @@ public class RoomRepository
 
     public async Task<bool> UpdateRoomAsync(Guid id, RoomUpdateDTO Room)
     {
-        var room = await GetByIdAsync(id);
+        var room = await GetRoomAsync(id);
+
         if (room == null)
             return false;
+
         room.PricePerNight = Room.PricePerNight;
         room.HotelId = Room.HotelId;
+
         _context.Rooms.Update(room);
+
         return await _context.SaveChangesAsync() > 0;
     }
 }
